Tests fixed, renamed, updated, new added
