+++ LanguageFeatures/Patterns/type_inference_A20_t01.dart
@@ -4,107 +4,41 @@
 
 /// @assertion
 /// To type check a pattern p being matched against a value of type M:
-/// ...
-/// Relational:
 ///
-/// Let C be the static type of the right operand constant expression.
+/// Logical-or and logical-and: Type check each branch using M as the matched
+/// value type.
 ///
-/// If the operator is a comparison (<, <=, >, or >=), then it is a compile-time
-/// error if:
-///
-/// - M does not define that operator,
-/// - C is not assignable to the operator's parameter type,
-/// - or if the operator's return type is not assignable to bool.
-///
-/// Else the operator is == or !=. It is a compile-time error if C is not
-/// assignable to T? where T is M's == method parameter type.
-/// The language screens out null before calling the underlying == method, which
-/// is why T? is the allowed type. Since Object declares == to accept Object on
-/// the right, this compile-time error can only happen if a user-defined class
-/// has an override of == with a covariant parameter.
-///
-/// @description Check that it is a compile-time error if the operator is a
-/// comparison (<, <=, >, or >=) and M does not define that operator
+/// @description Check that if both branch of a logical-or pattern fail a type
+/// check in refutable context then there is no error but match fails
 /// @author sgrekhov22@gmail.com
 
 // SharedOptions=--enable-experiment=patterns
 
+import "../../Utils/expect.dart";
+
 String test1() {
-  String s = "";
-  switch(s) {
-    case < 0:
-//       ^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-      return "<";
-    case <= 0:
-//       ^^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-      return "<=";
-    case > 0:
-//       ^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-
-      return ">";
-    case >= 0:
-//       ^^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-
-      return ">=";
+  switch(<num>[]) {
+    case List<int> v1 || List<int> v1:
+      return "match";
     default:
-      return "default";
+      return "no match";
   }
 }
 
 String test2() {
-  String s = "";
-  return switch(s) {
-    < 0 => "<",
-//  ^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-    <= 0 => "<=",
-//  ^^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-    > 0 => ">",
-//  ^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-    >= 0 => ">=",
-//  ^^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-    _ => "default"
-  };
+  if (<num>[] case List<int> v1 || List<int> v1) {
+    return "match";
+  }
+  return "no match";
 }
 
-String test3() {
-  String s = "";
-  if (s case < 0) return "<";
-//           ^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-  if (s case <= 0) return "<=";
-//           ^^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-  if (s case > 0) return ">";
-//           ^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-  if (s case >= 0) return ">=";
-//           ^^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-  return "default";
-}
+String test3() => switch(<num>[]) {
+    List<int> v1 || List<int> v1 => "match",
+    _ => "no match"
+  };
 
 main() {
-  test1();
-  test2();
-  test3();
+  Expect.equals("no match", test1());
+  Expect.equals("no match", test2());
+  Expect.equals("no match", test3());
 }
