+++ LanguageFeatures/Patterns/type_inference_A21_t01.dart
@@ -5,62 +5,99 @@
 /// @assertion
 /// To type check a pattern p being matched against a value of type M:
 /// ...
-/// Cast:
+/// Relational:
 ///
-/// Resolve the type name to a type X. It is a compile-time error if the name
-/// does not refer to a type.
+/// Let C be the static type of the right operand constant expression.
 ///
-/// Type-check the subpattern using X as the matched value type.
+/// If the operator is a comparison (<, <=, >, or >=), then it is a compile-time
+/// error if:
 ///
-/// @description Check that it is a compile-time error if X does not refer to a
-/// type
+/// - M does not define that operator,
+/// - C is not assignable to the operator's parameter type,
+/// - or if the operator's return type is not assignable to bool.
+///
+/// Else the operator is == or !=. It is a compile-time error if C is not
+/// assignable to T? where T is M's == method parameter type.
+/// The language screens out null before calling the underlying == method, which
+/// is why T? is the allowed type. Since Object declares == to accept Object on
+/// the right, this compile-time error can only happen if a user-defined class
+/// has an override of == with a covariant parameter.
+///
+/// @description Check that it is a compile-time error if the operator is a
+/// comparison (<, <=, >, or >=) and M does not define that operator
 /// @author sgrekhov22@gmail.com
 
 // SharedOptions=--enable-experiment=patterns
 
-const INT = 42;
-
 String test1() {
-  num x = 0;
-  switch (x) {
-    case var y as INT:
-//                ^^^
+  String s = "";
+  switch(s) {
+    case < 0:
+//       ^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-      return "int";
-    case final z as "String":
-//                  ^^^^^^^^
+      return "<";
+    case <= 0:
+//       ^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-      return "String";
+      return "<=";
+    case > 0:
+//       ^^^
+// [analyzer] unspecified
+// [cfe] unspecified
+
+      return ">";
+    case >= 0:
+//       ^^^^
+// [analyzer] unspecified
+// [cfe] unspecified
+
+      return ">=";
     default:
       return "default";
   }
 }
 
 String test2() {
-  num x = 0;
-  return switch (x) {
-    var y as INT => "int",
-//           ^^^
+  String s = "";
+  return switch(s) {
+    < 0 => "<",
+//  ^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-    final z as "String" => "String",
-//             ^^^^^^^^
+    <= 0 => "<=",
+//  ^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-    _ => "default";
+    > 0 => ">",
+//  ^^^
+// [analyzer] unspecified
+// [cfe] unspecified
+    >= 0 => ">=",
+//  ^^^^
+// [analyzer] unspecified
+// [cfe] unspecified
+    _ => "default"
   };
 }
 
 String test3() {
-  num x = 0;
-  if (x case var y as INT) return "int";
-//                    ^^^
+  String s = "";
+  if (s case < 0) return "<";
+//           ^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-  if (x case final z as "String") return "String";
-//                      ^^^^^^^^
+  if (s case <= 0) return "<=";
+//           ^^^^
+// [analyzer] unspecified
+// [cfe] unspecified
+  if (s case > 0) return ">";
+//           ^^^
+// [analyzer] unspecified
+// [cfe] unspecified
+  if (s case >= 0) return ">=";
+//           ^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
   return "default";
