+++ LanguageFeatures/Patterns/type_inference_A26_t03.dart
@@ -16,8 +16,8 @@
 /// In a declaration context, the required type of p is M, as is the static type
 /// of the variable introduced by p.
 ///
-/// @description Check that in an assignment context, the required type of p is
-/// the unpromoted static type of the variable that p resolves to
+/// @description Check that in a declaration context, the required type of p is
+/// M, as is the static type of the variable introduced by p.
 /// @author sgrekhov22@gmail.com
 
 // SharedOptions=--enable-experiment=patterns,records
@@ -26,16 +26,11 @@ import "../../Utils/static_type_helper.dart";
 
 main() {
   num a = 0;
-  a = 42;
   a.expectStaticType<Exactly<num>>();
 
   var (num b,) = (0,);
-  (b,) = (42,);
   b.expectStaticType<Exactly<num>>();
 
   var c = 2 > 1 ? 42 : 3.14;
-  c = 0;
-  c.expectStaticType<Exactly<num>>();
-  c = 3.14;
   c.expectStaticType<Exactly<num>>();
 }
