+++ LanguageFeatures/Patterns/type_inference_A28_t01.dart
@@ -5,21 +5,27 @@
 /// @assertion
 /// To type check a pattern p being matched against a value of type M:
 /// ...
-/// Map:
-/// i. Calculate the value's entry key type K and value type V, and key context
-///     C:
-///   a. If p has type arguments <K, V> for some K and V then use those, and C
-///     is K.
-///   b. Else if M implements Map<K, V> for some K and V then use those, and C
-///     is K.
-///   c. Else if M is dynamic then K and V are dynamic and C is _.
-///   d. Else K and V are Object? and C is _.
-/// ii. Type-check each key expression using C as the context type.
-/// iii. Type-check each value subpattern using V as the matched value type.
-/// vi. The required type of p is Map<K, V>.
+/// List:
 ///
-/// @description Check that if p has a type arguments <K, V>, then value's value
-/// type is V
+/// i. Calculate the value's element type E:
+///
+///   a. If p has a type argument T, then E is the type T.
+///   b. Else if M implements List<T> for some T then E is T.
+///   c. Else if M is dynamic then E is dynamic.
+///   d. Else E is Object?.
+///
+/// ii. Type-check each non-rest element subpattern using E as the matched value
+///   type. Note that we calculate a single element type and use it for all
+///   subpatterns. In:
+///     var [a, b] = [1, 2.3];
+///   both a and b use num as their matched value type.
+///
+/// iii. If there is a matching rest element, type-check its subpattern using
+///   List<E> as the matched value type.
+///
+/// iv. The required type of p is List<E>.
+///
+/// @description Check that if p has a type argument T, then E is the type T.
 /// @author sgrekhov22@gmail.com
 
 // SharedOptions=--enable-experiment=patterns
@@ -28,10 +34,12 @@ import "../../Utils/expect.dart";
 import "../../Utils/static_type_helper.dart";
 
 String test1() {
-  switch ({"key1": 1}) {
-    case <Object, num>{"key1": var a}:
+  switch ([1, 2]) {
+    case <num>[var a, var b]:
       a.expectStaticType<Exactly<num>>();
+      b.expectStaticType<Exactly<num>>();
       a = 3.14;
+      b = 3.14;
       return "match";
     default:
       return "no match";
@@ -39,9 +47,11 @@ String test1() {
 }
 
 String test2() {
-  switch ({"key1": 1}) {
-    case <Object, num>{"key1": final a, ...}:
+  switch ([1, 2, 3]) {
+    case <num>[final a, final b, ...final c]:
       a.expectStaticType<Exactly<num>>();
+      b.expectStaticType<Exactly<num>>();
+      c.expectStaticType<Exactly<List<num>>>();
       return "match";
     default:
       return "no match";
@@ -49,37 +59,44 @@ String test2() {
 }
 
 String test3() {
-  if ({"key1": 1} case <Object, num>{"key1": var a}) {
+  if ([1, 2] case <num>[var a, var b]) {
     a.expectStaticType<Exactly<num>>();
+    b.expectStaticType<Exactly<num>>();
     a = 3.14;
+    b = 3.14;
     return "match";
   }
   return "no match";
 }
 
 String test4() {
-  if ({"key1": 1} case <Object, num>{"key1": final a, ...}) {
+  if ([1, 2, 3] case <num>[final a, final b, ...final c]) {
     a.expectStaticType<Exactly<num>>();
+    b.expectStaticType<Exactly<num>>();
+    c.expectStaticType<Exactly<List<num>>>();
     return "match";
   }
   return "no match";
 }
 
 String test5() =>
-  switch ({"key1": 1}) {
-    <Object, num>{"key1": var a} when
-        a.expectStaticType<Exactly<num>>() is num && (a = 3.14) > 0 => "match",
+  switch ([1, 2]) {
+    <num>[var a, var b] when
+        a.expectStaticType<Exactly<num>>() is num &&
+        b.expectStaticType<Exactly<num>>() is num => "match",
     _ => "no match"
   };
 
 String test6() =>
-  switch ({"key1": 1}) {
-    <Object, num>{"key1": final a, ...} when
-        a.expectStaticType<Exactly<num>>() is num => "match",
+  switch ([1, 2, 3]) {
+    <num>[final a, final b, ...final c] when
+        a.expectStaticType<Exactly<num>>() is num &&
+        b.expectStaticType<Exactly<num>>() is num &&
+        c.expectStaticType<Exactly<List<num>>>() is List<num> => "match",
     _ => "no match"
   };
 
-main() {/*
+main() {
   var <num>[a1, b1] = [1, 2];
   a1.expectStaticType<Exactly<num>>();
   b1.expectStaticType<Exactly<num>>();
@@ -95,5 +112,5 @@ main() {/*
   Expect.equals("match", test3());
   Expect.equals("match", test4());
   Expect.equals("match", test5());
-  Expect.equals("match", test6());*/
+  Expect.equals("match", test6());
 }
